Add support for master.cfg SSHSlave configuration.
This is a bit of a hack in how it works, the correct usage in master.cfg is:

  slave = SSHBuildSlaveParent(<usual settings>)
  ssh1 = SSHBuildSlave(slave, <name>, <username>, <host>, <sshslave values>,
                       <usual slave settings>)
  c['slaves'] = [slave, ssh1]

The sshslave can be used as a normal slave would be there is absolutely no
difference.

On the slave side, the sshslave config is attached to the builder: the slave is
not aware of what an sshslave is.  When the builder launches the right
RunProcess class will be called which will run the commands locally or remotely.
(There is no code for this yet)

File Upload/Download and LogFileWatchers are next.